Atualizacao automática

+++ src/main/java/br/org/coletivoJava/fw/erp/implementacao/contapagarreceber/json_bind_galax_pay/FaturaAssinatura/ItfDTOFaturaAssinatura.java
@@ -91,7 +91,7 @@ public interface ItfDTOFaturaAssinatura
     }
 
     @Override
-    public default int getId() {
-        return (int) getValorPorReflexao();
+    public default Long getId() {
+        return (Long) getValorPorReflexao();
     }
 }
